# Gpl 473 performance improvements