--- a/.github/workflows/ci_docker.yml
+++ b/.github/workflows/ci_docker.yml
@@ -5,6 +5,9 @@ on:
     branches:
       - develop
       - master
+    tags:
+      - test-bug-*
+      
     paths-ignore:
       - "README.md"
 env: