--- a/.github/workflows/ci_docker.yml
+++ b/.github/workflows/ci_docker.yml
@@ -5,7 +5,6 @@ on:
     branches:
       - develop
       - master
-      
     paths-ignore:
       - "README.md"
 env: