Closes #54