@harrietc52 2020-07-14
Performance Improvements lighthouse POST samples_declarations
4 min 25 sec => 1 min 3 sec

Before (with errors)
20000  = 36 seconds

find duplicates =  8, 9, 8, 9
find non exist = 5, 5, 5, 4, seconds
add_errors = 45, 10 seconds
after pre, before post = 23, 23, 37, 26, 25 seconds
after post = 19 sec

70000 = 5 mins

find duplicates = 2 mins
find_non_exist_samples = 1 mins 16 sec
add_errors 	70000=4 minutes, 500=19sec
after pre, before post = 1 min 20 seconds, 1
start post =25 sec



After (with errors)
20000 = 24, 32 seconds

find duplicates = 1 sec
find non exist = 1 sec
add_errors = 17 sec (20000=10, 9, 8, 9, 18, 16 secs)
after pre, before post = 18sec, 22


70000 = 1 min 22 sec

find duplicates = 1 sec
find non exist = 1 sec
add_errors    	70000=8 mins, 500=18, 19 sec
after pre, before post = 


After (without errors)
20000 = 25 seconds

find duplicates = 1
find non exist = 1
add_errors = 0
after pre, before post = 24 sec

After (without errors)
70000 = 

find duplicates = 1
find non exist = 2
add_errors = 0
after pre, before post = 24 sec